README - temporarily independent, idealistic noob
This is my first attempt at coding a project myself, not simply following a brief tutorial or primarily using another person's code. I just barely started teaching myself to program in Python and to use Blender, so everything is likely very rough, at least for now. 

I will be separating the project into components/features as if it was a group project in order to save things properly, stay organized and improve my workflow. 

Normally, I'd do this privately, but I cannot afford a paid version, so it's public. If you found me, you're welcome to make gentle suggestions. I'm always lookiong to improve. However, I'm not currently looking for full-on collaboration (yet). 

I want to try to make as much of this game by myself as I can before I try to bring anyone else into it, even for free, so I learn more about the process and flesh out my plans for the project. I want to have a more solidified idea of what I want it to become before I ask for overall input or collaboration--besides minor coding help since I'm a noob.